backend/API_GatewayWS/requirements.txt
fastapi 
uvicorn 
httpx 
python-jose 
python-dotenv